PDFBOX-5889: support long values for COSInteger
git-svn-id: https://svn.apache.org/repos/asf/pdfbox/branches/3.0@1921497 13f79535-47bb-0310-9956-ffa450edef68

+++ debugger/src/main/java/org/apache/pdfbox/debugger/PDFDebugger.java
@@ -1237,7 +1237,7 @@ public class PDFDebugger extends JFrame implements Callable<Integer>, HyperlinkL
         }
         if (selectedNode instanceof COSInteger)
         {
-            return Integer.toString(((COSInteger) selectedNode).intValue());
+            return Long.toString(((COSInteger) selectedNode).longValue());
         }
         if (selectedNode instanceof COSName)
         {
